% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coletar_pronunciamentos_intervalo.R
\name{coletar_pronunciamentos_intervalo}
\alias{coletar_pronunciamentos_intervalo}
\title{Coleta pronunciamentos da ALMG em um intervalo de datas}
\usage{
coletar_pronunciamentos_intervalo(ini, fim, por_pagina = 100)
}
\arguments{
\item{ini}{Data inicial da consulta no formato "AAAAMMDD". Ex: "19950101".}

\item{fim}{Data final da consulta no formato "AAAAMMDD". Ex: "19951231".}

\item{por_pagina}{Número de registros por página da API. Máximo recomendado: 100.}
}
\value{
Um \code{tibble} contendo os pronunciamentos no período com as colunas:
\describe{
\item{controle}{Número de controle do pronunciamento}
\item{autor}{Nome do parlamentar}
\item{partido}{Sigla do partido}
\item{tipo}{Tipo do pronunciamento (Discurso, Questão de Ordem, etc.)}
\item{resumo}{Resumo textual}
\item{assunto}{Assuntos indexados}
\item{data}{Data do pronunciamento}
}
}
\description{
Consulta a API de dados abertos da Assembleia Legislativa de Minas Gerais (ALMG)
e retorna uma tabela com os metadados de pronunciamentos realizados no período especificado.
Os resultados são paginados automaticamente.
}
\examples{
\dontrun{
# Coletar pronunciamentos do ano de 2000
dados_2000 <- coletar_pronunciamentos_intervalo("20000101", "20001231")
}

}
