% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_autores_almg_intervalo.R
\name{get_autores_almg_intervalo}
\alias{get_autores_almg_intervalo}
\title{Baixa e empilha autores das proposições da ALMG para um intervalo de anos}
\usage{
get_autores_almg_intervalo(anos, tipo = "CSV")
}
\arguments{
\item{anos}{Vetor numérico com os anos desejados, ex: 2015:2023}

\item{tipo}{Tipo de arquivo para download (padrão: "CSV")}
}
\value{
Tibble com dados dos autores de proposições da ALMG.
}
\description{
Esta função baixa os dados de proposições da ALMG para cada ano no intervalo informado,
extrai e "explode" os autores das proposições (um autor por linha),
retornando um tibble com dados dos autores e identificação da proposição.
}
\examples{
\dontrun{
autores <- get_autores_almg_intervalo(2015:2023)
}
}
