% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bancadas.R
\name{get_bancadas}
\alias{get_bancadas}
\title{Coleta lista de Bancadas de Representação Partidária da ALMG}
\usage{
get_bancadas(formato = "json")
}
\arguments{
\item{formato}{String. Formato da resposta, "json" (padrão) ou "xml".}
}
\value{
Um tibble com informações das bancadas, ou tibble vazio em caso de erro.
}
\description{
Consulta a API da Assembleia Legislativa de Minas Gerais para obter a lista
de bancadas de representação partidária da legislatura atual.
}
\examples{
\dontrun{
bancadas <- get_bancadas()
}

}
