% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deputados_em_exercicio.R
\name{get_deputados_em_exercicio}
\alias{get_deputados_em_exercicio}
\title{Lista os deputados em exercício na legislatura atual da ALMG}
\usage{
get_deputados_em_exercicio(formato = "json")
}
\arguments{
\item{formato}{Formato da resposta da API, pode ser "json" (padrão) ou "xml".}
}
\value{
Um \code{tibble} com as colunas:
\describe{
\item{id}{Número identificador do deputado}
\item{nome}{Nome completo do deputado}
\item{partido}{Sigla do partido político}
\item{temDetalhe}{Indica se o deputado possui detalhes adicionais (boolean)}
\item{tagLocalizacao}{Código da localização do deputado}
\item{cargoRepresentacaoPartidaria}{Cargo de representação partidária (se existir)}
\item{liderancas}{Lista de lideranças associadas ao deputado (se existir)}
\item{dataCargoRepresentacaoPartidaria}{Data do cargo de representação partidária (se existir)}
\item{cargoMesa}{Cargo na mesa diretora (se existir)}
\item{dataCargoMesa}{Data do cargo na mesa diretora (se existir)}
\item{sexo}{Sexo do deputado}
}
}
\description{
Consulta a API de dados abertos da Assembleia Legislativa de Minas Gerais (ALMG)
e retorna uma tabela com informações dos deputados em exercício.
}
\examples{
\dontrun{
deputados <- get_deputados_em_exercicio()
print(deputados)
}

}
