% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_partidos.R
\name{get_partidos}
\alias{get_partidos}
\title{Coleta informações das Partidos parlamentares da ALMG}
\usage{
get_partidos(formato = "json")
}
\arguments{
\item{formato}{String. Formato da resposta: "json" (padrão) ou "xml".}
}
\value{
Um tibble com as informações das Partidos, ou tibble vazio em caso de erro.
}
\description{
Consulta a API da Assembleia Legislativa de Minas Gerais para obter dados das Partidos parlamentares
da legislatura atual.
}
\examples{
\dontrun{
partidos <- get_partidos()
}
}
