% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pronunciamento_por_num_cont.R
\name{get_pronunciamento_por_num_cont}
\alias{get_pronunciamento_por_num_cont}
\title{Coleta detalhes de um pronunciamento pelo número de controle}
\usage{
get_pronunciamento_por_num_cont(numCont, formato = "json")
}
\arguments{
\item{numCont}{Integer. Número de controle do pronunciamento.}

\item{formato}{String. Formato da resposta: "json" (padrão) ou "xml".}
}
\value{
Um tibble com os detalhes do pronunciamento, ou tibble vazio em caso de erro.
}
\description{
Consulta a API da Assembleia Legislativa de Minas Gerais para obter os detalhes de um pronunciamento específico identificado pelo número de controle.
}
\examples{
\dontrun{
pronunciamento <- get_pronunciamento_por_num_cont(3047)
}
}
