% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pronunciamento.R
\name{get_pronunciamentos}
\alias{get_pronunciamentos}
\title{Coleta detalhes de pronunciamentos dado um vetor de IDs}
\usage{
get_pronunciamentos(numCont_vec)
}
\arguments{
\item{numCont_vec}{Vetor de IDs dos pronunciamentos (numéricos ou caracteres)}
}
\value{
Tibble com os dados dos pronunciamentos e coluna texto limpa (texto sem HTML)
}
\description{
Coleta detalhes de pronunciamentos dado um vetor de IDs
}
