% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proposicoes_almg_intervalo.R
\name{get_proposicoes_almg_intervalo}
\alias{get_proposicoes_almg_intervalo}
\title{Obtem dados de proposições da ALMG para um intervalo de anos}
\usage{
get_proposicoes_almg_intervalo(anos, tipo = "CSV")
}
\arguments{
\item{anos}{Vetor numérico com os anos desejados.}

\item{tipo}{Tipo do arquivo para download ("CSV" por padrão).}
}
\value{
Tibble com dados das proposições.
}
\description{
Baixa dados de proposições da ALMG para cada ano no intervalo informado,
seleciona as colunas principais relacionadas à proposição e retorna um tibble.
}
\examples{
\dontrun{
# Baixar proposições dos anos de 2015 a 2023
proposicoes <- get_proposicoes_almg_intervalo(2015:2023)
head(proposicoes)
}
}
