% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proposicoes_completas_almg.R
\name{get_proposicoes_completas_almg}
\alias{get_proposicoes_completas_almg}
\title{Baixa proposições da ALMG com textos completos para um intervalo de anos}
\usage{
get_proposicoes_completas_almg(
  anos,
  tipo = "CSV",
  incluir_outros_textos = FALSE
)
}
\arguments{
\item{anos}{Vetor numérico com os anos desejados, ex: 2015:2023}

\item{tipo}{Tipo do arquivo para download (padrão: "CSV")}

\item{incluir_outros_textos}{Booleano, se TRUE inclui os textos adicionais (default FALSE)}
}
\value{
Tibble com as proposições e seus textos completos.
}
\description{
Esta função baixa os dados de proposições da ALMG para cada ano no intervalo informado,
baixa os textos (original e outros textos) de cada proposição e retorna um tibble
combinando as informações.
}
