% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_registro_deputados.R
\name{get_registro_deputados}
\alias{get_registro_deputados}
\title{Obtém registros detalhados de deputados pelo(s) ID(s)}
\usage{
get_registro_deputados(ids_deputados, formato = "json")
}
\arguments{
\item{ids_deputados}{Vetor numérico com um ou mais identificadores de deputados.}

\item{formato}{String. Formato da resposta: "json" (padrão) ou "xml".}
}
\value{
Um \code{tibble} com os principais campos dos registros dos deputados, incluindo lideranças desaninhadas.
Retorna tibble vazio em caso de erro ou resposta vazia.
}
\description{
Consulta a API da Assembleia Legislativa de Minas Gerais para retornar os dados detalhados de um ou mais deputados
pelo(s) seu(s) identificador(es) único(s).
}
\examples{
\dontrun{
registros <- get_registro_deputados(c(9673, 7752, 28859))
}

}
